% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqFunctions.R
\name{getFreqTable}
\alias{getFreqTable}
\title{Get Stylo Frequency List Function
This function performs a stylometric analysis of your corpus and creates a file called "table_with_frequencies"
in your working directory. The file is imported as dataframe. Please name the output according to specified settings
e.g. styloFreqList_c_1.}
\usage{
getFreqTable(
  type = "w",
  size = 1,
  path,
  corpusfoldername = "corpus",
  mfw.min = 100,
  mfw.max = 500,
  mfw.incr = 100,
  culling.min = 20,
  culling.max = 70,
  culling.incr = 20,
  mfw.list.cutoff = 5000,
  distance.measure = "eder",
  output = freq_list
)
}
\arguments{
\item{type}{Do you want to create word ("w") or chacacter ("c") n-grams?}

\item{size}{Indicate the name of the corpus folder. Default is "corpus".}

\item{path}{Indicate the path of your folder containing the corpus folder.}
}
\description{
Get Stylo Frequency List Function
This function performs a stylometric analysis of your corpus and creates a file called "table_with_frequencies"
in your working directory. The file is imported as dataframe. Please name the output according to specified settings
e.g. styloFreqList_c_1.
}
\examples{
getFreqTable()

}
